+++ src/Eccube/Service/PurchaseFlow/ProcessResult.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Eccube\Service;
+namespace Eccube\Service\PurchaseFlow;
 
 
 class ProcessResult
