--- tests/Eccube/Tests/Service/ValidatableItemHolderProcessorTest.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: chihiro_adachi
- * Date: 2017/07/03
- * Time: 15:47
- */
-
-namespace Eccube\Tests\Service;
-
-
-class ValidatableItemHolderProcessorTest extends \PHPUnit_Framework_TestCase
-{
-    public function testHoge()
-    {
-
-    }
-}
